feat: 프로젝트 설정

+++ DS2024/Division.cs
@@ -4,7 +4,7 @@
     {
         public static decimal Divide(int dividend, int divisor)
         {
-            decimal quotient = dividend / divisor;
+            decimal quotient = (decimal)dividend / divisor;
             return quotient;
         }
     }
